# backend/openapi.yaml
openapi: 3.0.3
info:
  title: FinTrak
  description: |
    Personal finance tracker API. All API endpoints are prefixed with `/api/v1`.

    ## Authentication
    Protected routes require a Bearer token in the `Authorization` header:
    ```
    Authorization: Bearer <token>
    ```

    Tokens are obtained via `/auth/register` or `/auth/login`.
  version: 0.1.0-alpha.0
  contact:
    name: FinTrak
  license:
    name: AGPL v3.0
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
paths:
...
  /transactions:
    get:
      tags:
        - Transactions
      summary: List transactions
      description: Get transactions with optional filtering, sorting, and pagination
      operationId: getTransactions
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: accountId
          schema:
            type: string
            format: uuid
          description: Filter by account ID
        - in: query
          name: categoryId
          schema:
            type: string
            format: uuid
          description: Filter by category ID
        - in: query
          name: payeeId
          schema:
            type: string
            format: uuid
          description: Filter by payee ID
        - in: query
          name: search
          schema:
            type: string
          description: Search in description (case-insensitive)
        - in: query
          name: dateFrom
          schema:
            type: string
            format: date
          description: Filter transactions from this date (YYYY-MM-DD)
        - in: query
          name: dateTo
          schema:
            type: string
            format: date
          description: Filter transactions until this date (YYYY-MM-DD)
        - in: query
          name: type
          schema:
            type: string
            enum: [debit, credit]
          description: Filter by transaction type
        - in: query
          name: amount
          schema:
            type: number
          description: Filter by exact amount
        - in: query
          name: uncategorized
          schema:
            type: string
            enum: ["true", "false"]
          description: Filter for uncategorized transactions only
        - in: query
          name: linked
          schema:
            type: string
            enum: ["true", "false"]
          description: Filter by link status
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [date, amount, createdAt]
            default: date
          description: Sort column
        - in: query
          name: sortOrder
          schema:
            type: string
            enum: [ASC, DESC]
            default: DESC
          description: Sort direction
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
          description: Page number (1-indexed)
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
          description: Items per page (0 = no pagination, return all)
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags:
        - Transactions
      summary: Create a new transaction
      operationId: createTransaction
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionRequest"
      responses:
        "201":
          description: Transaction created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /transactions/{id}:
    patch:
      tags:
        - Transactions
      summary: Update a transaction
      description: Update transaction fields (category, tags, notes, payee, date, description, amount, type, billing cycle)
      operationId: updateTransaction
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Transaction ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTransactionRequest"
      responses:
        "200":
          description: Transaction updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      tags:
        - Transactions
      summary: Delete a transaction
      operationId: deleteTransaction
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Transaction ID
      responses:
        "200":
          description: Transaction deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  schemas:
...
    CreateTransactionRequest:
      type: object
      required:
        - accountId
        - date
        - description
        - amount
        - type
      properties:
        accountId:
          type: string
          format: uuid
        date:
          type: string
          format: date
        description:
          type: string
        amount:
          type: number
          minimum: 0
          exclusiveMinimum: true
        type:
          type: string
          enum: [debit, credit]
        categoryId:
          type: string
          format: uuid
        payeeId:
          type: string
          format: uuid
        tags:
          type: array
          items:
            type: string
        notes:
          type: string
        billingCycleId:
          type: string
          format: uuid
    UpdateTransactionRequest:
      type: object
      properties:
        categoryId:
          type: string
          format: uuid
          nullable: true
        tags:
          type: array
          items:
            type: string
        notes:
          type: string
        payeeId:
          type: string
          format: uuid
          nullable: true
        date:
          type: string
          format: date
        description:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [debit, credit]
        accountId:
          type: string
          format: uuid
        billingCycleId:
          type: string
          format: uuid
          nullable: true
...
    Transaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        accountId:
          type: string
          format: uuid
        accountName:
          type: string
        date:
          type: string
          format: date-time
        description:
          type: string
        amount:
          type: number
          format: double
        type:
          type: string
          enum: [debit, credit]
        categoryId:
          type: string
          format: uuid
        categoryName:
          type: string
        categoryIcon:
          type: string
        categoryColor:
          type: string
        payeeId:
          type: string
          format: uuid
        payee:
          type: string
        tags:
          type: array
          items:
            type: string
        notes:
          type: string
        isLinked:
          type: boolean
        linkCount:
          type: integer
        linkId:
          type: string
          format: uuid
        billingCycleId:
          type: string
          format: uuid
        billingCycleLabel:
          type: string
        loanAccountId:
          type: string
          format: uuid
          description: Loan / EMI account this transaction is attached to as an EMI payment
        loanAccountName:
          type: string
        createdAt:
          type: string
          format: date-time
        isSummary:
          type: boolean
    TransactionListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        pages:
          type: integer
...
    ErrorResponse:
      type: object
      required:
        - errors
      properties:
        errors:
          type: array
          minItems: 1
          items:
            type: object
            required:
              - message
            properties:
              field:
                type: string
              tag:
                type: string
              message:
                type: string
    MessageResponse:
      type: object
      properties:
        message:
          type: string
  responses:
    BadRequest:
      description: Bad request (invalid input)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: User does not have permission to access this resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/register or /auth/login